+++ hardware/EnOceanESP3.h
@@ -1,13 +1,13 @@
 #pragma once
 
-#include "EnOceanESP.h"
+#include "EnOceanEEP.h"
 #include "ASyncSerial.h"
 #include "DomoticzHardware.h"
 #include <map>
 
 #define ENOCEAN3_READ_BUFFER_SIZE 65 * 1024
 
-class CEnOceanESP3 : public EnOceanESP, public AsyncSerial, public CDomoticzHardwareBase
+class CEnOceanESP3 : public CEnOceanEEP, public AsyncSerial, public CDomoticzHardwareBase
 {
 	enum _eEnOcean_Receive_State
 	{
@@ -47,6 +47,24 @@ class CEnOceanESP3 : public EnOceanESP, public AsyncSerial, public CDomoticzHard
 	void readCallback(const char *data, size_t len);
 
 	void ReloadVLDNodes();
+
+ 	const char *GetPacketTypeLabel(const uint8_t PT);
+	const char *GetPacketTypeDescription(const uint8_t PT);
+
+	const char *GetReturnCodeLabel(const uint8_t RC);
+	const char *GetReturnCodeDescription(const uint8_t RC);
+
+	const char *GetFunctionReturnCodeLabel(const uint8_t RC);
+	const char *GetFunctionReturnCodeDescription(const uint8_t RC);
+
+	const char *GetEventCodeLabel(const uint8_t EC);
+	const char *GetEventCodeDescription(const uint8_t EC);
+
+	const char *GetCommonCommandLabel(const uint8_t CC);
+	const char *GetCommonCommandDescription(const uint8_t CC);
+
+	const char *GetSmarkAckCodeLabel(const uint8_t SA);
+	const char *GetSmartAckCodeDescription(const uint8_t SA);
 
       private:
 	_eEnOcean_Receive_State m_receivestate;
